Move mergeIncrementalData to internal
