+++ src/core/QueryInfo.ts
@@ -5,8 +5,11 @@ import type { ApolloCache, Cache } from "@apollo/client/cache";
 import type { FetchResult } from "@apollo/client/link";
 import type { Unmasked } from "@apollo/client/masking";
 import { graphQLResultHasError } from "@apollo/client/utilities";
-import { mergeIncrementalData } from "@apollo/client/utilities";
-import { DeepMerger, isNonEmptyArray } from "@apollo/client/utilities/internal";
+import {
+  DeepMerger,
+  isNonEmptyArray,
+  mergeIncrementalData,
+} from "@apollo/client/utilities/internal";
 
 import type { ObservableQuery } from "./ObservableQuery.js";
 import type { QueryManager } from "./QueryManager.js";
