+++ src/utilities/internal/index.ts
@@ -33,6 +33,7 @@ export { makeReference } from "./makeReference.js";
 export { maybeDeepFreeze } from "./maybeDeepFreeze.js";
 export { mergeDeep } from "./mergeDeep.js";
 export { mergeDeepArray } from "./mergeDeepArray.js";
+export { mergeIncrementalData } from "./mergeIncrementalData.js";
 export { mergeOptions } from "./mergeOptions.js";
 export { onAnyEvent } from "./onAnyEvent.js";
 export { preventUnhandledRejection } from "./preventUnhandledRejection.js";
